查询缴费支付完成

+++ shared-trad-customer-serv/shared-trad-customer-api/src/main/java/cn/suparking/customer/api/beans/parkfee/ParkingOrder.java
@@ -35,7 +35,7 @@ public class ParkingOrder {
 
     private Integer aggregatedMaxAmount;
 
-    private Integer parkingMinutes;
+    private Long parkingMinutes;
 
     private DiscountInfo discountInfo;
